set up authentication and modularize views
