+++ stock_portfolio/stock_portfolio/__init__.py
@@ -8,5 +8,6 @@ def main(global_config, **settings):
     config.include('pyramid_jinja2')
     config.include('.models')
     config.include('.routes')
+    config.include('.security')
     config.scan()
     return config.make_wsgi_app()
